Use bower and grunt for building webapp
This removes all library code from the repo. Instead it uses bower for
dependency management and grunt for building the webapp inside the
_site folder. There is also a grunt task which allows live editing.
Simply start with grunt and open the preview in the browser. Any
change in any source file will automatically cause a reload of the
page.

Usage:
```
cd web/
npm install
grunt
```